Renamed all giphy name references inside ui package to only gif
