+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -40,7 +40,7 @@ import org.wordpress.android.ui.comments.EditCommentActivity;
 import org.wordpress.android.ui.domains.DomainRegistrationActivity;
 import org.wordpress.android.ui.domains.DomainRegistrationDetailsFragment;
 import org.wordpress.android.ui.domains.DomainSuggestionsFragment;
-import org.wordpress.android.ui.giphy.GiphyPickerActivity;
+import org.wordpress.android.ui.gif.GifPickerActivity;
 import org.wordpress.android.ui.history.HistoryAdapter;
 import org.wordpress.android.ui.history.HistoryDetailContainerFragment;
 import org.wordpress.android.ui.main.AddContentAdapter;
@@ -438,7 +438,7 @@ public interface AppComponent extends AndroidInjector<WordPress> {
 
     void inject(JetpackRemoteInstallFragment jetpackRemoteInstallFragment);
 
-    void inject(GiphyPickerActivity object);
+    void inject(GifPickerActivity object);
 
     void inject(PlansListAdapter object);
