+++ WordPress/src/main/java/org/wordpress/android/ui/ActivityLauncher.java
@@ -38,7 +38,7 @@ import org.wordpress.android.ui.activitylog.list.ActivityLogListActivity;
 import org.wordpress.android.ui.comments.CommentsActivity;
 import org.wordpress.android.ui.domains.DomainRegistrationActivity;
 import org.wordpress.android.ui.domains.DomainRegistrationActivity.DomainRegistrationPurpose;
-import org.wordpress.android.ui.giphy.GiphyPickerActivity;
+import org.wordpress.android.ui.gif.GifPickerActivity;
 import org.wordpress.android.ui.history.HistoryDetailActivity;
 import org.wordpress.android.ui.history.HistoryDetailContainerFragment;
 import org.wordpress.android.ui.history.HistoryListItem.Revision;
@@ -179,7 +179,7 @@ public class ActivityLauncher {
         properties.put("from", activity.getClass().getSimpleName());
         AnalyticsTracker.track(Stat.GIF_PICKER_ACCESSED, properties);
 
-        Intent intent = new Intent(activity, GiphyPickerActivity.class);
+        Intent intent = new Intent(activity, GifPickerActivity.class);
         intent.putExtra(WordPress.SITE, site);
         activity.startActivityForResult(intent, requestCode);
     }
